requirements_dev.txt
pytest==3.4.2
pytest-runner==2.11.1
tox==2.9.1
cookiecutter>=1.4.0
pytest-cookies==0.3.0
watchdog==0.8.3

pip==10.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
Sphinx==1.7.5
sphinx-rtd-theme==0.2.4
twine==1.10.0

mock==2.0.0
numpydoc==0.7.0
nose==1.3.7
